src/requirements.txt
pandas
fastapi
gunicorn
uvicorn[standard]
sqlalchemy
pydantic
alembic
python-dotenv
gradio
tenacity
openai
jsons
pydantic
httpx
fastapi
IPython
redis
chromadb